Move files to relevant directories
This is so they are mapped accurately with what's in src/.
