+++ tests/conftest.py
@@ -3,9 +3,9 @@ import pytest
 
 pytest_plugins = [
     "tests.fixtures.utils.fixture_helper_synonym",
-    "tests.fixtures.utils.fixture_preprocess",
-    "tests.fixtures.utils.fixture_helper_embedding",
-    "tests.fixtures.utils.fixture_helper_intent_uis",
+    "tests.fixtures.make_data.fixture_preprocess",
+    "tests.fixtures.make_features.fixture_helper_embedding",
+    "tests.fixtures.make_features.fixture_helper_intent_uis",
     "tests.fixtures.make_features.fixture_helper_bert",
 ]
